+++ client/CHANGELOG.md
@@ -3,12 +3,25 @@
 All notable changes to the "vscode-abap-remote-fs" extension will be documented in this file.
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
+
+## Unreleased
+
+### Changed
+
+- inline open original in diff view
+
+### Fixed
+
+- lock restore based on all documents rather than visible ones
+
 ## [0.6.4] 2019-04-25
+
 ### Fixed
 
 - fix open sapgui in windows
 
 ## [0.6.3] 2019-04-24
+
 ### Fixed
 
 - retry lock on change