+++ client/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
-## Unreleased
+## [0.6.5] 2019-04-25
 
 ### Added
 
@@ -14,6 +14,7 @@ Format based on [Keep a Changelog](http://keepachangelog.com/)
 
 - inline open original in diff view
 - default to stateless connections
+- debounce object locking
 - queue all stateful calls
 
 ### Fixed